feat: Update NumberBox to winui3/release/1.7.1
